pyproject.toml
[project]
name = "chatroom"
version = "0.1.0"
description = "Chatroom app with Flask-AppBuilder + OAuth2"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "flask-appbuilder==3.4.5",
    "flask-sqlalchemy==2.5.1",
    "SQLAlchemy==1.4.39",
    "Flask==2.1.3",
    "python-dotenv>=1.1.1",
    "bcrypt>=4.0.0",
    "flask-jwt-extended>=4.6.0",
    "authlib==1.3.0",
    "requests>=2.32.4",
]
